receiving_emails/requirements.txt
kafka-python
Flask
requests
